move to vite-vue-pwa
