+++ examples/vite-vue-pwa/README.md
