+++ examples/vite-vue-pwa/index.html
