+++ examples/vite-vue-pwa/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vite-vue-pwa-spinner",
+  "name": "vite-vue-pwa",
   "private": true,
   "version": "0.0.0",
   "type": "module",
