+++ examples/vite-vue-pwa/src/components/CachedImageWidget.vue
