+++ examples/vite-vue-pwa/src/components/LoadingSpinner.vue
