+++ examples/vite-vue-pwa/src/components/RandomColorSWAPIWidget.vue
