+++ examples/vite-vue-pwa/src/components/ServiceWorkerWidget.vue
