+++ examples/vite-vue-pwa/src/components/SharedArrayBufferWidget.vue
