+++ examples/vite-vue-pwa/src/index.ts
