+++ examples/vite-vue-pwa/src/sw.ts
