+++ examples/vite-vue-pwa/vite.config.ts
